Only point the clock background once
This now paints the clock background ones to an image surface during startup.
Later, this image can just be used and the three clock hands have to be added.

This fixes a memory leak where previously we (implicitely) created several
surfaces on each clock refresh due to the push_group() calls. It took too long
for the garbage collector to collect these surfaces which meant a high memory
usage.

Signed-off-by: Uli Schlachter <psychon@znc.in>